980 - Merge EWS Data Viewer into PopupChart
